chore: refactor gulpfile into individual files
We take advantage of gulp's ability to break build scripts into multiple
files to ease the maintenance of the build scripts.  Some positive
benefits of this include:
 1. No longer compiling/minifying for development builds
 2. Karma is able to read from static file instead of parsing a JSON to
 figure out where to look.
 3. Karma is in headless mode.
 4. Only transfer server files that we need, instead of everything.
