+++ karma.conf.js
@@ -1,7 +1,5 @@
 // Karma configuration
 // Generated on Wed Mar 30 2016 15:50:28 GMT+0100 (WAT)
-// Import manifest JSON from latest client build
-const manifest = require('./bin/client/rev-manifest');
 
 module.exports = (config) => {
   config.set({
@@ -15,10 +13,10 @@ module.exports = (config) => {
 
     // list of files / patterns to load in the browser
     files : [
-      `bin/client/${manifest['js/vendor/vendor.min.js']}`,
+      `bin/client/js/vendor/vendor.min.js`,
       'node_modules/angular-mocks/angular-mocks.js',
       'test/client-unit/mocks/*.js',
-      `bin/client/${manifest['js/bhima/bhima.min.js']}`,
+      `bin/client/js/bhima/bhima.min.js`,
       'bin/client/modules/**/*.html',
       { pattern : 'bin/client/i18n/locale/*.js', included : false, served : true },
       'test/client-unit/**/*.spec.js',
@@ -66,7 +64,7 @@ module.exports = (config) => {
 
     // start these browsers
     // available browser launchers : https://npmjs.org/browse/keyword/karma-launcher
-    browsers : ['Chrome'],
+    browsers : ['ChromeHeadless'],
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
